+++ NEWS.md
@@ -2,6 +2,8 @@
 - Fixed an issue where sub-elements of a port with `'converse` record
   mode view would have the wrong direction when the port was partially
   associated.
+- Package instantiation declarations are now allowed in the entity
+  declarative part (#1266).
 - Several other minor bugs were resolved (#1257, #1260, #1262, #1264).
 
 ## Version 1.17.1 - 2025-07-26